refactoring project structure and improving design
