+++ project_database/.env
@@ -1,0 +1,5 @@
+USER="postgres"
+PASSWORD=1234
+POSTGRES_DB= "db"
+PORT=5432
+HOST="localhost"
