+++ project_database/requirements.txt
@@ -1,0 +1,8 @@
+python-dotenv
+
+numpy
+pandas
+psycopg2-binary
+pyparsing
+PySocks
+os
